pyproject.toml
[tool.poetry]
name = "typer-config"
version = "0.1.2"
description = "Utilities for working with configuration files in typer CLIs. "
authors = ["Matt Anderson <matt@manderscience.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "typer_config"}]
repository = "https://github.com/maxb2/typer-config"
keywords = ["typer"]

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0"
toml = {version = "^0.10.2", optional = true}
pyyaml = {version = "^6.0", optional = true}

[tool.poetry.extras]
toml = ["toml"]
yaml = ["pyyaml"]
all = ["toml", "pyyaml"]

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.263"
pylint = "^2.17.3"
black = "^23.3.0"
isort = "^5.12.0"
pytest = "^7.3.1"
mypy = "^1.2.0"
types-toml = "^0.10.8.6"
types-pyyaml = "^6.0.12.9"
coverage = "^7.2.5"

[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.1.9"
mkdocstrings = {extras = ["python"], version = "^0.21.2"}
mike = "^1.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
